requirements.txt
# Core dependencies for the RAG pipeline
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.23.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
